+++ src/design/plone/contenttypes/tests/test_ct_news.py
@@ -3,6 +3,7 @@ from design.plone.contenttypes.testing import (
     DESIGN_PLONE_CONTENTTYPES_INTEGRATION_TESTING,
     DESIGN_PLONE_CONTENTTYPES_API_FUNCTIONAL_TESTING,
 )
+from design.plone.contenttypes.interfaces import IDesignPloneContentType
 from plone import api
 from plone.app.testing import setRoles
 from plone.app.testing import SITE_OWNER_NAME
@@ -60,6 +61,11 @@ class TestNews(unittest.TestCase):
             sorted(("Image", "File", "Link", "Document")),
             sorted(portal_types["News Item"].allowed_content_types),
         )
+
+    def test_news_provide_design_pct_marker_interface(self):
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
+        news = api.content.create(container=self.portal, type="News Item", title="News")
+        self.assertTrue(IDesignPloneContentType.providedBy(news))
 
 
 class TestNewsApi(unittest.TestCase):