+++ src/design/plone/contenttypes/upgrades/upgrades.py
@@ -1025,3 +1025,19 @@ def to_5600(context):
         if not getattr(venue, "exclude_from_nav", None):
             setattr(venue, "exclude_from_nav", False)
             venue.reindexObject(idxs=["exclude_from_nav"])
+
+
+def to_5700(context):
+    logger.info("Reindex Event and News to fix SearchableText indexing issue")
+
+    brains = api.content.find(portal_type=["Event", "News Item"])
+    tot = len(brains)
+    logger.info("Found {} documents.".format(tot))
+    i = 0
+    for brain in brains:
+        i += 1
+        if i % 100 == 0:
+            logger.info("Progress: {}/{}".format(i, tot))
+        doc = brain.getObject()
+        doc.reindexObject(idxs=["SearchableText"])
+    logger.info("Ends of reindex")